pyproject.toml
[tool.poetry]
name = "km-backend"
version = "0.1.0"
description = "kali magazine back-end"
authors = ["Minsu Cho <ethanminsu8969@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
django-environ = "^0.11.2"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
